Move aC call-arg analyzer into its own core plugin ##plugins
The aC/aCe/aCf family of commands lives in cmd_anal.inc.c next to
hundreds of unrelated analysis commands, which makes it hard to iterate
on and obscures the fact that the call-argument analyzer is a cohesive
feature with a single job. Extract it into libr/core/p/core_callargs.c
as a standalone core plugin that intercepts the aC command prefix.

The plugin keeps the exact behaviour of the previous implementation:
sentinel priming of A0..A7 before emulation so untouched arg registers
are reported as unknown, shortest-path ESIL via abpe with a BB-window
fallback for huge or unanalysed functions, 32-bit sign-extension
cleanup on raw values, and plaintext, JSON and r2-command output modes
for both aCe and aCf variants. All helper functions that used to live
in cmd_anal.inc.c (ace_prime_regs, ace_run_emulation, ace_append_arg,
ace_append_raw, cmd_anal_aCe_at, cmd_anal_aCf_iter, cmd_anal_aC) now
live inside the plugin and are hidden behind a single ace_handle entry
point invoked from the RCorePlugin .call callback.

The plugin registers via the usual core plugin channels: an entry in
libr/config.h (generated from dist/plugins-cfg/plugins.def.cfg), the
matching callargs.mk build fragment, a meson core_plugins entry and
an extern declaration in r_core.h. The dead case 'C' dispatch and the
help_msg_aC definition are removed from cmd_anal.inc.c.
