+++ libr/core/cmd_anal.inc.c
@@ -157,17 +157,6 @@ static RCoreHelpMessage help_msg_avg = {
 	NULL
 };
 
-static RCoreHelpMessage help_msg_aC = {
-	"Usage:", "aC[fej*] [addr-of-call]", " # analyze function call args via ESIL emulation",
-	"aCe", " [addr]", "resolve args of the call at [addr] (uses 'abpe' to emulate the shortest path)",
-	"aCej", " [addr]", "same as aCe but in JSON (one object per argument with raw + typed value)",
-	"aCe*", " [addr]", "emit r2 commands to set a CCu comment with the resolved call args",
-	"aCf", "", "apply aCe to every call instruction in the current function",
-	"aCfj", "", "apply aCej to every call in the current function (JSON array)",
-	"aCf*", "", "apply aCe* to every call in the current function (sets comments at each call site)",
-	NULL
-};
-
 static RCoreHelpMessage help_msg_aaf = {
 	"Usage:", "aaf[efrt?]", " # analyse all function, see also 'af' and 'afna'",
 	"aaf", "", "same as afr@@c:isq",
@@ -15495,426 +15484,6 @@ static void cmd_anal_classes(RCore *core, const char *input) {
 	}
 }
 
-// Output mode for aCe: plaintext summary, structured JSON or r2 comment commands.
-typedef enum {
-	ACE_MODE_PLAIN = 0,
-	ACE_MODE_JSON,
-	ACE_MODE_R2,
-} AceMode;
-
-// Sentinel written into arg registers before emulation. Any register that
-// still holds this value afterwards has not been touched by the path
-// emulation, so the argument is reported as "unknown" rather than as 0.
-// The value is arbitrary but deliberately avoids small integers and common
-// pointer patterns so that simple arithmetic on SP won't accidentally
-// produce it.
-#define ACE_SENTINEL  ((ut64)0xdeadbeefcafebabeULL)
-
-// Upper bound on the number of register slots we watch with a sentinel. It
-// covers the widest common calling conventions (amd64 = 6, aarch64 = 8).
-#define ACE_MAX_REG_ARGS 8
-
-// Above this basic-block count the shortest-path emulation becomes too
-// expensive, so we fall back to emulating just the enclosing basic block.
-#define ACE_MAX_FCN_BBS 256
-
-// True if `v` equals the sentinel, meaning the path emulation never wrote
-// to this register.
-static inline bool ace_is_unknown(ut64 v) {
-	return v == ACE_SENTINEL;
-}
-
-// Return a "displayable" value for `v`. Strips 32-bit sign extension that
-// the ESIL engine often carries around on 64-bit architectures (e.g.
-// 0xffffffff80001000 -> 0x80001000) so that rendered values match the
-// literal the code loaded.
-static ut64 ace_display_value(ut64 v) {
-	if ((v >> 32) == 0xffffffffULL && (v & 0x80000000ULL)) {
-		return v & 0xffffffffULL;
-	}
-	return v;
-}
-
-// Reset the ESIL register file (ar0) while preserving SP, then seed the
-// arg-register aliases A0..A{nregs-1} with a sentinel. Callers run this
-// immediately before driving emulation so that untouched registers stay
-// recognisable after the run.
-static void ace_prime_regs(RCore *core, int nregs) {
-	// Ensure ESIL and its stack memory exist; both commands are idempotent.
-	r_core_cmd0 (core, "aei");
-	r_core_cmd0 (core, "aeim");
-	ut64 sp = r_reg_getv (core->anal->reg, "SP");
-	r_reg_arena_zero (core->anal->reg);
-	if (sp != 0 && sp != UT64_MAX) {
-		r_reg_setv (core->anal->reg, "SP", sp);
-	}
-	if (nregs > ACE_MAX_REG_ARGS) {
-		nregs = ACE_MAX_REG_ARGS;
-	}
-	int i;
-	char name[16];
-	for (i = 0; i < nregs; i++) {
-		snprintf (name, sizeof (name), "A%d", i);
-		r_reg_setv (core->anal->reg, name, ACE_SENTINEL);
-	}
-}
-
-// Pick an emulation strategy and execute it. For functions whose CFG is
-// small enough we rely on `abpe` (shortest-path ESIL from function entry),
-// otherwise we emulate only within the enclosing basic block - that keeps
-// the command responsive on very large functions or on code outside any
-// analysed function.
-static void ace_run_emulation(RCore *core, ut64 pcv) {
-	RAnalFunction *fcn = r_anal_get_fcn_in (core->anal, pcv, -1);
-	if (fcn && r_list_length (fcn->bbs) <= ACE_MAX_FCN_BBS) {
-		r_core_cmdf (core, "abpe 0x%08"PFMT64x, pcv);
-		return;
-	}
-	RList *bbs = r_anal_get_blocks_in (core->anal, pcv);
-	RAnalBlock *bb = bbs ? (RAnalBlock *)r_list_first (bbs) : NULL;
-	if (bb) {
-		r_core_cmdf (core, "aepc 0x%08"PFMT64x, bb->addr);
-		r_core_cmdf (core, "aesou 0x%08"PFMT64x, pcv);
-	}
-	r_list_free (bbs);
-}
-
-// Append a formatted argument (typed value + raw numeric) to either a JSON
-// array being built in `pj` or a plaintext buffer `sb`. Exactly one of pj/sb
-// must be non-NULL. Arguments whose backing register still holds the
-// sentinel are reported as unknown instead of being printed as 0.
-static void ace_append_arg(RCore *core, RAnalFuncArg *arg, bool on_stack, int idx, PJ *pj, RStrBuf *sb) {
-	bool unknown = ace_is_unknown (arg->src);
-	ut64 disp = ace_display_value (arg->src);
-	char *typed = NULL;
-	if (!unknown && arg->fmt) {
-		typed = print_fcn_arg (core, arg->orig_c_type, arg->name, arg->fmt, arg->src, on_stack, 0);
-		if (typed) {
-			r_str_trim (typed);
-			if (R_STR_ISEMPTY (typed)) {
-				R_FREE (typed);
-			}
-		}
-	}
-	if (pj) {
-		pj_o (pj);
-		pj_kn (pj, "index", idx);
-		if (R_STR_ISNOTEMPTY (arg->name)) {
-			pj_ks (pj, "name", arg->name);
-		}
-		if (arg->orig_c_type) {
-			pj_ks (pj, "type", arg->orig_c_type);
-		} else if (arg->c_type) {
-			pj_ks (pj, "type", arg->c_type);
-		}
-		if (R_STR_ISNOTEMPTY (arg->cc_source)) {
-			pj_ks (pj, "source", arg->cc_source);
-		}
-		pj_kn (pj, "size", arg->size);
-		pj_kb (pj, "known", !unknown);
-		if (!unknown) {
-			pj_kn (pj, "raw", disp);
-			if (typed) {
-				pj_ks (pj, "value", typed);
-			}
-		}
-		pj_end (pj);
-	} else {
-		if (idx > 0) {
-			r_strbuf_append (sb, ", ");
-		}
-		if (R_STR_ISNOTEMPTY (arg->name)) {
-			r_strbuf_appendf (sb, "%s=", arg->name);
-		}
-		if (unknown) {
-			r_strbuf_append (sb, "?");
-		} else if (typed) {
-			r_strbuf_appendf (sb, "%s /*0x%"PFMT64x"*/", typed, disp);
-		} else {
-			r_strbuf_appendf (sb, "0x%"PFMT64x, disp);
-		}
-	}
-	free (typed);
-}
-
-// Append raw register-based arguments (A0..A{n-1}) when no signature is
-// known. Sentinel-valued registers are rendered as "?" / "known":false.
-static void ace_append_raw(RCore *core, int nargs, PJ *pj, RStrBuf *sb) {
-	if (nargs < 0) {
-		nargs = DEFAULT_NARGS;
-	}
-	if (nargs > ACE_MAX_REG_ARGS) {
-		nargs = ACE_MAX_REG_ARGS;
-	}
-	int i;
-	char regname[16];
-	for (i = 0; i < nargs; i++) {
-		snprintf (regname, sizeof (regname), "A%d", i);
-		ut64 v = r_reg_getv (core->anal->reg, regname);
-		bool unknown = ace_is_unknown (v);
-		ut64 disp = ace_display_value (v);
-		if (pj) {
-			pj_o (pj);
-			pj_kn (pj, "index", i);
-			pj_ks (pj, "source", regname);
-			pj_kb (pj, "known", !unknown);
-			if (!unknown) {
-				pj_kn (pj, "raw", disp);
-			}
-			pj_end (pj);
-		} else {
-			if (i > 0) {
-				r_strbuf_append (sb, ", ");
-			}
-			if (unknown) {
-				r_strbuf_appendf (sb, "%s=?", regname);
-			} else {
-				r_strbuf_appendf (sb, "%s=0x%"PFMT64x, regname, disp);
-			}
-		}
-	}
-}
-
-// Core routine: prime the ESIL register file with a sentinel, emulate the
-// shortest path (or the enclosing basic block as a fallback) to `pcv`, and
-// render the arguments of the call found there in the requested mode.
-static void cmd_anal_aCe_at(RCore *core, ut64 pcv, AceMode mode) {
-	// Prime A0..A{MAX-1} with a sentinel so we can tell afterwards which
-	// arg registers the emulation actually wrote to.
-	ace_prime_regs (core, ACE_MAX_REG_ARGS);
-	// Emulate from function entry (or enclosing BB for large functions).
-	ace_run_emulation (core, pcv);
-
-	RAnalOp *op = r_core_anal_op (core, pcv, -1);
-	if (!op) {
-		return;
-	}
-	bool is_call = (op->type == R_ANAL_OP_TYPE_CALL
-			|| op->type == R_ANAL_OP_TYPE_UCALL
-			|| op->type == R_ANAL_OP_TYPE_ICALL
-			|| op->type == R_ANAL_OP_TYPE_RCALL
-			|| op->type == R_ANAL_OP_TYPE_CCALL
-			|| op->type == R_ANAL_OP_TYPE_UCCALL);
-
-	// Resolve the callee: prefer a real RAnalFunction at the call target,
-	// otherwise fall back to a matching flag.
-	const char *fcn_name = NULL;
-	RAnalFunction *callee = NULL;
-	if (is_call && op->jump != UT64_MAX && op->jump != 0) {
-		callee = r_anal_get_function_at (core->anal, op->jump);
-		if (callee) {
-			fcn_name = callee->name;
-		} else {
-			RFlagItem *item = r_flag_get_in (core->flags, op->jump);
-			if (item) {
-				fcn_name = item->name;
-			}
-		}
-	}
-	char *key = fcn_name ? r_type_func_name (core->anal->sdb_types, fcn_name) : NULL;
-	const char *ret_type = key ? r_type_func_ret (core->anal->sdb_types, key) : NULL;
-	const char *cc_name = NULL;
-	if (key) {
-		cc_name = r_anal_cc_func (core->anal, key);
-	}
-	if (!cc_name) {
-		cc_name = r_anal_cc_default (core->anal);
-	}
-
-	// Temporarily bump SP to match carg.c stack-arg conventions.
-	int s_width = (core->anal->config->bits == 64) ? 8 : 4;
-	ut64 spv = r_reg_getv (core->anal->reg, "SP");
-	r_reg_setv (core->anal->reg, "SP", spv + s_width);
-	RList *list = fcn_name ? r_core_get_func_args (core, fcn_name) : NULL;
-
-	PJ *pj = NULL;
-	RStrBuf *sb = NULL;
-	if (mode == ACE_MODE_JSON) {
-		pj = r_core_pj_new (core);
-		pj_o (pj);
-		pj_kn (pj, "addr", pcv);
-		pj_ks (pj, "op", is_call ? "call" : r_anal_optype_tostring (op->type));
-		if (is_call && op->jump != UT64_MAX) {
-			pj_kn (pj, "target", op->jump);
-		}
-		if (fcn_name) {
-			pj_ks (pj, "function", fcn_name);
-		}
-		if (ret_type) {
-			pj_ks (pj, "return_type", ret_type);
-		}
-		if (cc_name) {
-			pj_ks (pj, "calling_convention", cc_name);
-		}
-		pj_ka (pj, "args");
-	} else {
-		sb = r_strbuf_new ("");
-		if (fcn_name) {
-			if (ret_type) {
-				r_strbuf_appendf (sb, "%s%s%s(", ret_type,
-					(ret_type[strlen (ret_type) - 1] == '*') ? "" : " ",
-					fcn_name);
-			} else {
-				r_strbuf_appendf (sb, "%s(", fcn_name);
-			}
-		} else if (is_call && op->jump != UT64_MAX && op->jump != 0) {
-			r_strbuf_appendf (sb, "0x%08"PFMT64x"(", op->jump);
-		} else {
-			r_strbuf_append (sb, "?(");
-		}
-	}
-
-	bool has_typed_args = list && !r_list_empty (list);
-	if (has_typed_args) {
-		bool on_stack = false;
-		RAnalFuncArg *first = r_list_first (list);
-		if (first && first->cc_source && r_str_startswith (first->cc_source, "stack")) {
-			on_stack = true;
-		}
-		int idx = 0;
-		RListIter *iter;
-		RAnalFuncArg *arg;
-		r_list_foreach (list, iter, arg) {
-			ace_append_arg (core, arg, on_stack, idx, pj, sb);
-			idx++;
-		}
-	} else {
-		// No signature found: fall back to the raw A0..An register values.
-		int nargs = DEFAULT_NARGS;
-		if (callee) {
-			int n = r_anal_var_count_args (callee);
-			if (n > 0) {
-				nargs = n;
-			}
-		}
-		ace_append_raw (core, nargs, pj, sb);
-	}
-	r_list_free (list);
-	r_reg_setv (core->anal->reg, "SP", spv);
-	free (key);
-
-	if (mode == ACE_MODE_JSON) {
-		pj_end (pj); // args
-		pj_end (pj); // root
-		char *j = pj_drain (pj);
-		r_cons_printf (core->cons, "%s\n", j);
-		free (j);
-	} else {
-		r_strbuf_append (sb, ")");
-		char *s = r_strbuf_drain (sb);
-		if (mode == ACE_MODE_R2) {
-			char *u = (char *)r_base64_encode_dyn ((const ut8 *)s, -1);
-			if (u) {
-				r_cons_printf (core->cons, "'@0x%08"PFMT64x"'CCu base64:%s\n", pcv, u);
-				free (u);
-			}
-		} else {
-			r_cons_println (core->cons, s);
-		}
-		free (s);
-	}
-	r_anal_op_free (op);
-}
-
-// Enumerate every call instruction in the function enclosing core->addr and
-// invoke cmd_anal_aCe_at on each, wrapping the output appropriately for JSON
-// or r2-command modes.
-static void cmd_anal_aCf_iter(RCore *core, AceMode mode) {
-	RAnalFunction *fcn = r_anal_get_fcn_in (core->anal, core->addr, 0);
-	if (!fcn) {
-		R_LOG_ERROR ("No function at 0x%08"PFMT64x, core->addr);
-		return;
-	}
-	RList *calls = r_list_new ();
-	if (!calls) {
-		return;
-	}
-	RListIter *it;
-	RAnalBlock *bb;
-	r_list_foreach (fcn->bbs, it, bb) {
-		ut64 at = bb->addr;
-		while (at < bb->addr + bb->size) {
-			RAnalOp *op = r_core_anal_op (core, at, R_ARCH_OP_MASK_BASIC);
-			if (!op || op->size < 1) {
-				r_anal_op_free (op);
-				break;
-			}
-			int t = op->type;
-			if (t == R_ANAL_OP_TYPE_CALL || t == R_ANAL_OP_TYPE_UCALL
-				|| t == R_ANAL_OP_TYPE_ICALL || t == R_ANAL_OP_TYPE_RCALL
-				|| t == R_ANAL_OP_TYPE_CCALL || t == R_ANAL_OP_TYPE_UCCALL) {
-				ut64 *p = R_NEW (ut64);
-				if (p) {
-					*p = at;
-					r_list_append (calls, p);
-				}
-			}
-			at += op->size;
-			r_anal_op_free (op);
-		}
-	}
-	if (mode == ACE_MODE_JSON) {
-		r_cons_print (core->cons, "[");
-		RListIter *cit;
-		ut64 *pv;
-		bool first = true;
-		r_list_foreach (calls, cit, pv) {
-			if (!first) {
-				r_cons_print (core->cons, ",");
-			}
-			first = false;
-			cmd_anal_aCe_at (core, *pv, ACE_MODE_JSON);
-		}
-		r_cons_print (core->cons, "]\n");
-	} else {
-		RListIter *cit;
-		ut64 *pv;
-		r_list_foreach (calls, cit, pv) {
-			cmd_anal_aCe_at (core, *pv, mode);
-		}
-	}
-	r_list_free (calls);
-}
-
-// ripped from disasm.c: dupe code from there
-static void cmd_anal_aC(RCore *core, const char *input) {
-	// aCf / aCfj / aCf*: iterate all calls in the current function.
-	if (*input == 'f') {
-		AceMode mode = ACE_MODE_PLAIN;
-		if (input[1] == 'j') {
-			mode = ACE_MODE_JSON;
-		} else if (input[1] == '*') {
-			mode = ACE_MODE_R2;
-		} else if (input[1] != '\0' && input[1] != ' ') {
-			r_core_cmd_help (core, help_msg_aC);
-			return;
-		}
-		cmd_anal_aCf_iter (core, mode);
-		return;
-	}
-	if (*input != 'e') {
-		r_core_cmd_help (core, help_msg_aC);
-		return;
-	}
-	input++; // skip 'e'
-	AceMode mode = ACE_MODE_PLAIN;
-	if (*input == 'j') {
-		mode = ACE_MODE_JSON;
-		input++;
-	} else if (*input == '*') {
-		mode = ACE_MODE_R2;
-		input++;
-	}
-	while (*input == ' ') {
-		input++;
-	}
-	ut64 pcv = R_STR_ISEMPTY (input) ? core->addr : r_num_math (core->num, input);
-	if (pcv == 0 || pcv == UT64_MAX) {
-		r_core_cmd_help (core, help_msg_aC);
-		return;
-	}
-	cmd_anal_aCe_at (core, pcv, mode);
-}
 
 static bool core_anal_abf(RCore *core, const char* input) {
 	if (strchr (input, '?')) {
@@ -16486,9 +16055,6 @@ static int cmd_anal(void *data, const char *input) {
 		break;
 	case 'c': // "ac"
 		cmd_anal_classes (core, input + 1);
-		break;
-	case 'C': // "aC"
-		cmd_anal_aC (core, input + 1);
 		break;
 	case 'i': // "ai"
 	       	cmd_anal_info (core, input + 1);
